pyproject.toml
[tool.poetry]
name = "tazboard"
version = "0.0.1"
description = "An internal application for monitoring and management of the taz website"
authors = ["Lyudmila Vaseva <lyudmila@ctrl.alt.coop>",
           "Jonatan Zint <jonatan@ctrl.alt.coop>" ]
license = "GPL-2.0-or-later"
include = [
    "tazboard/frontend/templates/**/*",
    "tazboard/frontend/static/**/*"
]
exclude = [
    "tazboard/frontend/static_src/**/"
]

[tool.poetry.dependencies]
python = "^3.6.1"
django = "^3.0.0"
dj-database-url = "^0.5.0"
psycopg2-binary = "^2.8.5"
whitenoise = "^5.1.0"
elasticsearch = "^7.0.0"
requests = "^2.24.0"
djangorestframework = "^3.11.0"
pyyaml = "^5.3.1"
django-filter = "^2.3.0"
uritemplate = "^3.0.1"

[tool.poetry.scripts]
full-build = 'build:build'

[tool.poetry.dev-dependencies]
pytest-django = "^3.9.0"
ipdb = "^0.13.2"
pylint = "^2.4.4"
coverage = "^5.1"
flake8 = "^3.8.3"

[build-system]
requires = [
    "setuptools >= 35.0.2",
    "setuptools_scm >= 2.0.0, <3"
]
build-backend = "setuptools.build_meta"

